requirements.txt
grpcio==1.25.0
grpcio-tools==1.25.0
joblib==0.14.0
numpy==1.17.4
pandas==0.25.3
protobuf==3.11.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
